Langchain_AI_Agent/requirements.txt
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-tavily>=0.1.0
langgraph>=0.1.0
pyowm>=3.3.0
python-dotenv>=1.0.0
